Relax the byte channel interface requirement in FileChannelDataSource. Rename to SeekableByteChannelDataSource. Keep FileChannelDataSource for the sake of backward compatibility
